Service.Lighting/requirements.txt
flask
flask-cors
aioflask
requests
gevent
fastapi

flask-pymongo
pymongo

python-icc
